refactor: workspace専用CLI4件を削除する (#4915)
