+++ pyproject.toml
@@ -47,9 +47,7 @@ yt-benchmark-comments = "youtube_automation.entrypoints:yt_benchmark_comments"
 yt-bulk-update-desc = "youtube_automation.entrypoints:yt_bulk_update_desc"
 yt-bulk-update-synthetic-media = "youtube_automation.entrypoints:yt_bulk_update_synthetic_media"
 yt-captions-upload = "youtube_automation.entrypoints:yt_captions_upload"
-yt-channel = "youtube_automation.entrypoints:yt_channel"
 yt-channel-export = "youtube_automation.entrypoints:yt_channel_export"
-yt-channel-import = "youtube_automation.entrypoints:yt_channel_import"
 yt-channel-init = "youtube_automation.entrypoints:yt_channel_init"
 yt-channel-seed = "youtube_automation.entrypoints:yt_channel_seed"
 yt-channel-settings = "youtube_automation.entrypoints:yt_channel_settings"
@@ -57,14 +55,12 @@ yt-channel-status = "youtube_automation.entrypoints:yt_channel_status"
 yt-channels = "youtube_automation.entrypoints:yt_channels"
 yt-session-start = "youtube_automation.entrypoints:yt_session_start"
 yt-workflow-state = "youtube_automation.entrypoints:yt_workflow_state"
-yt-workspace-status = "youtube_automation.entrypoints:yt_workspace_status"
 yt-workflow-status = "youtube_automation.entrypoints:yt_workflow_status"
 yt-hybrid-runner = "youtube_automation.entrypoints:yt_hybrid_runner"
 yt-human-tasks = "youtube_automation.entrypoints:yt_human_tasks"
 yt-codex-canary-notify = "youtube_automation.entrypoints:yt_codex_canary_notify"
 yt-post-publish-state = "youtube_automation.entrypoints:yt_post_publish_state"
 yt-media-acceptance = "youtube_automation.entrypoints:yt_media_acceptance"
-yt-workspace-guard = "youtube_automation.entrypoints:yt_workspace_guard"
 yt-channel-trend = "youtube_automation.entrypoints:yt_channel_trend"
 yt-collection-preflight = "youtube_automation.entrypoints:yt_collection_preflight"
 yt-collection-serve = "youtube_automation.entrypoints:yt_collection_serve"
