+++ tests/repo/test_cli_harness_gate.py
@@ -25,13 +25,10 @@ LEGACY_CLI_ALLOWLIST: frozenset[str] = frozenset(
         "youtube_automation.commands.analytics.video_analyze",
         "youtube_automation.commands.analytics.vpd_rank",
         "youtube_automation.commands.analytics.win_pattern",
-        "youtube_automation.commands.channel.channel",
-        "youtube_automation.commands.channel.channel_import",
         "youtube_automation.commands.channel.channel_init",
         "youtube_automation.commands.channel.channel_seed",
         "youtube_automation.commands.channel.channel_settings",
         "youtube_automation.commands.channel.channel_status",
-        "youtube_automation.commands.channel.workspace_guard",
         "youtube_automation.commands.collections.collection_preflight",
         "youtube_automation.commands.collections.collection_serve",
         "youtube_automation.commands.collections.init_collection",
