+++ tests/repo/test_skill_api_call_estimate_contract.py
@@ -84,15 +84,11 @@ NON_BILLED_CLIS: dict[str, str] = {
     "yt-audio-visualizer-fill": "ローカルの Pillow 画像生成のみ",
     "yt-audio-studio": "ローカル音源の一覧表示・HTTP Range 配信のみ",
     "yt-automation-update": "バージョン pin bump / git 操作のみ",
-    "yt-channel": "ローカルの workspace channel 列挙のみ",
     "yt-channel-export": "ローカルの workspace channel コピーと config 検証のみ",
     "yt-channels": "ローカルの channel registry と Git pin の読み取りのみ",
     "yt-session-start": "ローカルの git 状態確認と yt-automation-update への委譲のみ",
-    "yt-channel-import": "ローカルの channel repository コピーと config 検証のみ",
     "yt-channel-init": "ローカルの config 雛形生成のみ",
-    "yt-workspace-guard": "ローカルの workspace チャンネル境界判定のみ",
     "yt-workflow-state": "ローカルの workflow-state.json 読み書きのみ",
-    "yt-workspace-status": "YouTube Data API 無料枠の channels.list 1 request のみ",
     "yt-workflow-status": "ローカルの workflow-state.json と実成果物の読み取り・HTML表示のみ",
     "yt-channel-trend": "収集済み analytics_data_*.json のローカル分析のみ",
     "yt-collection-preflight": "ローカルの事前チェックのみ",
