+++ tests/test_cli_stdio.py
@@ -25,17 +25,13 @@ EXPECTED_ENTRYPOINT_MODULES = {
     "yt-bulk-update-synthetic-media": "youtube_automation.commands.metadata.bulk_update_synthetic_media",
     "yt-captions-upload": "youtube_automation.commands.youtube.captions_upload",
     "yt-changelog-compile": "youtube_automation.commands.system.changelog_compile",
-    "yt-channel": "youtube_automation.commands.channel.channel",
     "yt-channel-export": "youtube_automation.commands.channel.channel_export",
-    "yt-channel-import": "youtube_automation.commands.channel.channel_import",
     "yt-channel-init": "youtube_automation.commands.channel.channel_init",
     "yt-channel-seed": "youtube_automation.commands.channel.channel_seed",
     "yt-channel-settings": "youtube_automation.commands.channel.channel_settings",
     "yt-channel-status": "youtube_automation.commands.channel.channel_status",
     "yt-channels": "youtube_automation.commands.system.channels",
     "yt-session-start": "youtube_automation.commands.system.session_start",
-    "yt-workspace-status": "youtube_automation.commands.channel.workspace_status",
-    "yt-workspace-guard": "youtube_automation.commands.channel.workspace_guard",
     "yt-workflow-state": "youtube_automation.commands.collections.workflow_state_cli",
     "yt-workflow-status": "youtube_automation.commands.collections.workflow_status",
     "yt-channel-trend": "youtube_automation.commands.analytics.channel_trend",
